fix salary bug

+++ employeeTwo.c
@@ -41,7 +41,7 @@ static int compareEmployeePhone(const void *targetPtr, PtrToConstEmployee tableV
 
 static int compareEmployeeSalary(const void *targetPtr, PtrToConstEmployee tableValuePtr)
 {
-    return * (long *) targetPtr != tableValuePtr->salary;
+    return * (double *) targetPtr != tableValuePtr->salary;
 }
 
 //These are called wrappers. These functions are what you will use in your main!!!
